+++ include/galois/filters/convolution.h
@@ -11,15 +11,15 @@ namespace gs {
     private:
         SP_Signal<T> in_signal = nullptr;
         SP_Signal<T> out_signal = nullptr;
-        int num_rows = 0;
-        int num_columns = 0;
-        int in_channels = 0;
-        int out_channels = 0;
-        int kernel_rows = 0;
-        int kernel_columns = 0;
+        size_t num_rows = 0;
+        size_t num_columns = 0;
+        size_t in_channels = 0;
+        size_t out_channels = 0;
+        size_t kernel_rows = 0;
+        size_t kernel_columns = 0;
 
-        // int padding_w = 0;
-        // int padding_h = 0;
+        // size_t padding_w = 0;
+        // size_t padding_h = 0;
 
         SP_NArray<T> w = nullptr;
         SP_NArray<T> b = nullptr;
@@ -30,13 +30,13 @@ namespace gs {
         Convolution(const bool for_clone_or_share) {}
         Convolution(const Convolution&) = delete;
         Convolution& operator=(const Convolution&) = delete;
-        Convolution(int num_rows, int num_columns, int in_channels, int out_channels, int kernel_w, int kernel_h);
+        Convolution(size_t num_rows, size_t num_columns, size_t in_channels, size_t out_channels, size_t kernel_w, size_t kernel_h);
 
         SP_Filter<T> share() override;
         SP_Filter<T> clone() override;
 
         void install_signals(const vector<SP_Signal<T>> &in_signals, const vector<SP_Signal<T>> &out_signals) override;
-        void set_dims(int batch_size) override;
+        void set_dims(size_t batch_size) override;
         void reopaque() override;
 
         vector<SP_NArray<T>> get_params() override;
