+++ include/galois/filters/embedding.h
@@ -11,8 +11,8 @@ namespace gs {
     private:
         SP_Signal<T> in_signal = nullptr;
         SP_Signal<T> out_signal = nullptr;
-        int in_size = 0;
-        int out_size = 0;
+        size_t in_size = 0;
+        size_t out_size = 0;
 
         SP_NArray<T> w = nullptr;
         SP_NArray<T> dw = nullptr;
@@ -21,13 +21,13 @@ namespace gs {
         Embedding(const bool for_share_or_clone) {}
         Embedding(const Embedding&) = delete;
         Embedding& operator=(const Embedding&) = delete;
-        Embedding(int in_size, int out_size);
+        Embedding(size_t in_size, size_t out_size);
 
         SP_Filter<T> share() override;
         SP_Filter<T> clone() override;
 
         void install_signals(const vector<SP_Signal<T>> &in_signals, const vector<SP_Signal<T>> &out_signals) override;
-        void set_dims(int batch_size) override;
+        void set_dims(size_t batch_size) override;
         void reopaque() override;
 
         vector<SP_NArray<T>> get_params() override;
