+++ include/galois/filters/max_pooling.h
@@ -11,26 +11,26 @@ namespace gs {
     private:
         SP_Signal<T> in_signal = nullptr;
         SP_Signal<T> out_signal = nullptr;
-        int kernel_rows = 0;
-        int kernel_columns = 0;
-        int stride_rows = 0;
-        int stride_columns = 0;
+        size_t kernel_rows = 0;
+        size_t kernel_columns = 0;
+        size_t stride_rows = 0;
+        size_t stride_columns = 0;
 
         // the following are set in set_dims
-        int num_rows = 0;
-        int num_columns = 0;
-        int channels = 0;
+        size_t num_rows = 0;
+        size_t num_columns = 0;
+        size_t channels = 0;
         SP_NArray<T> max_indexes = nullptr;
 
     public:
         MaxPooling(const MaxPooling&) = delete;
         MaxPooling& operator=(const MaxPooling) = delete;
-        MaxPooling(int kernel_rows, int kernel_columns, int stride_rows, int stride_columns);
+        MaxPooling(size_t kernel_rows, size_t kernel_columns, size_t stride_rows, size_t stride_columns);
 
         SP_Filter<T> share() override;
 
         void install_signals(const vector<SP_Signal<T>> &in_signals, const vector<SP_Signal<T>> &out_signals) override;
-        void set_dims(int batch_size) override;
+        void set_dims(size_t batch_size) override;
         void reopaque() override {}
 
         void forward() override;
